+++ src/main/java/com/getindata/kafka/connect/iceberg/sink/IcebergUtil.java
@@ -13,6 +13,7 @@ import org.apache.iceberg.catalog.Catalog;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.data.GenericAppenderFactory;
 import org.apache.iceberg.exceptions.NoSuchTableException;
+import org.apache.iceberg.relocated.com.google.common.collect.Sets;
 import org.apache.iceberg.relocated.com.google.common.primitives.Ints;
 import org.apache.iceberg.types.Types;
 import org.slf4j.Logger;
@@ -20,8 +21,8 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Locale;
 import java.util.Optional;
-
-import static org.apache.iceberg.TableProperties.*;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * @author Ismail Simsek
@@ -29,8 +30,8 @@ import static org.apache.iceberg.TableProperties.*;
 public class IcebergUtil {
     protected static final Logger LOGGER = LoggerFactory.getLogger(IcebergUtil.class);
 
-  public static Table createIcebergTable(Catalog icebergCatalog, TableIdentifier tableIdentifier,
-                                         Schema schema, Optional<String> partitionField) {
+    public static Table createIcebergTable(Catalog icebergCatalog, TableIdentifier tableIdentifier,
+                                           Schema schema, Optional<String> partitionField) {
 
         LOGGER.info("Creating table:'{}'\nschema:{}\nrowIdentifier:{}", tableIdentifier, schema,
                 schema.identifierFieldNames());
@@ -56,7 +57,7 @@ public class IcebergUtil {
         }).orElseGet(partitionBuilder::build);
 
         return icebergCatalog.buildTable(tableIdentifier, schema)
-                .withProperty(FORMAT_VERSION, "2")
+                .withProperty(TableProperties.FORMAT_VERSION, "2")
                 .withSortOrder(IcebergUtil.getIdentifierFieldsAsSortOrder(schema))
                 .withPartitionSpec(ps)
                 .create();
@@ -82,17 +83,35 @@ public class IcebergUtil {
     }
 
     public static FileFormat getTableFileFormat(Table icebergTable) {
-        String formatAsString = icebergTable.properties().getOrDefault(DEFAULT_FILE_FORMAT, DEFAULT_FILE_FORMAT_DEFAULT);
+        String formatAsString = icebergTable.properties().getOrDefault(
+                TableProperties.DEFAULT_FILE_FORMAT,
+                TableProperties.DEFAULT_FILE_FORMAT_DEFAULT);
         return FileFormat.valueOf(formatAsString.toUpperCase(Locale.ROOT));
     }
 
-    public static GenericAppenderFactory getTableAppender(Table icebergTable) {
+    public static long getTargetDataFileSize(Table table) {
+        return Optional.ofNullable(table.properties().get(TableProperties.WRITE_TARGET_FILE_SIZE_BYTES))
+                .map(Long::parseLong)
+                .orElse(TableProperties.WRITE_TARGET_FILE_SIZE_BYTES_DEFAULT);
+    }
+
+    public static GenericAppenderFactory getTableAppender(Table icebergTable, boolean upsert) {
         return new GenericAppenderFactory(
                 icebergTable.schema(),
                 icebergTable.spec(),
-                Ints.toArray(icebergTable.schema().identifierFieldIds()),
+                getEqualityFieldIds(icebergTable.spec(), icebergTable.schema(), upsert).stream().mapToInt(Number::intValue).toArray(),
                 icebergTable.schema(),
                 null);
+    }
+
+    public static Set<Integer> getEqualityFieldIds(PartitionSpec spec, Schema schema, boolean upsert) {
+        var equalityFieldIds = Sets.newHashSet(schema.identifierFieldIds());
+        if (upsert) {
+            // Note: https://iceberg.apache.org/docs/latest/flink/#upsert
+            // i am not sure why this is recommended, but just incase follow the docs ;)
+            equalityFieldIds.addAll(spec.fields().stream().map(PartitionField::sourceId).collect(Collectors.toSet()));
+        }
+        return equalityFieldIds;
     }
 
 }