+++ src/main/java/com/getindata/kafka/connect/iceberg/sink/tableoperator/BaseDeltaTaskWriter.java
@@ -1,34 +1,32 @@
 package com.getindata.kafka.connect.iceberg.sink.tableoperator;
 
-import org.apache.iceberg.FileFormat;
-import org.apache.iceberg.PartitionKey;
-import org.apache.iceberg.PartitionSpec;
-import org.apache.iceberg.Schema;
-import org.apache.iceberg.StructLike;
+import org.apache.iceberg.*;
 import org.apache.iceberg.data.InternalRecordWrapper;
 import org.apache.iceberg.data.Record;
 import org.apache.iceberg.io.BaseTaskWriter;
 import org.apache.iceberg.io.FileAppenderFactory;
 import org.apache.iceberg.io.FileIO;
 import org.apache.iceberg.io.OutputFileFactory;
-import org.apache.iceberg.relocated.com.google.common.collect.Sets;
 import org.apache.iceberg.types.TypeUtil;
 
 import java.io.IOException;
-import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 
-import static com.getindata.kafka.connect.iceberg.sink.tableoperator.CdcOperation.CREATE;
 import static com.getindata.kafka.connect.iceberg.sink.tableoperator.CdcOperation.DELETE;
 
 public abstract class BaseDeltaTaskWriter extends BaseTaskWriter<Record> {
 
     private final Schema schema;
     private final Schema deleteSchema;
+
     private final InternalRecordWrapper wrapper;
 
     private final InternalRecordWrapper keyWrapper;
+
     private final boolean upsert;
     private final boolean upsertKeepDeletes;
+    private final String cdcOpField;
 
     public BaseDeltaTaskWriter(PartitionSpec spec,
                                FileFormat format,
@@ -37,19 +35,21 @@ public abstract class BaseDeltaTaskWriter extends BaseTaskWriter<Record> {
                                FileIO io,
                                long targetFileSize,
                                Schema schema,
-                               List<Integer> equalityFieldIds,
+                               Set<Integer> equalityFieldIds,
                                boolean upsert,
-                               boolean upsertKeepDeletes) {
+                               boolean upsertKeepDeletes,
+                               String cdcOpField) {
         super(spec, format, appenderFactory, fileFactory, io, targetFileSize);
         this.schema = schema;
-        this.deleteSchema = TypeUtil.select(schema, Sets.newHashSet(equalityFieldIds));
+        this.deleteSchema = TypeUtil.select(schema, equalityFieldIds);
         this.wrapper = new InternalRecordWrapper(schema.asStruct());
         this.keyWrapper = new InternalRecordWrapper(deleteSchema.asStruct());
         this.upsert = upsert;
         this.upsertKeepDeletes = upsertKeepDeletes;
+        this.cdcOpField = cdcOpField;
     }
 
-    abstract RowDataDeltaWriter route(Record row);
+    abstract RecordDeltaWriter route(Record row);
 
     InternalRecordWrapper wrapper() {
         return wrapper;
@@ -57,22 +57,30 @@ public abstract class BaseDeltaTaskWriter extends BaseTaskWriter<Record> {
 
     @Override
     public void write(Record row) throws IOException {
-        RowDataDeltaWriter writer = route(row);
-        if (upsert && !row.getField("__op").equals(CREATE.getCode())) {// anything which not an insert is upsert
-            writer.delete(row);
+        RecordDeltaWriter writer = route(row);
+
+        Optional<CdcOperation> cdcOp = Optional.ofNullable(row.getField(this.cdcOpField))
+                .map(Object::toString)
+                .map(CdcOperation::getByCode);
+
+        boolean isDelete = cdcOp.equals(Optional.of(DELETE));
+
+        if (upsert || isDelete) {
+            // only use deleteKey due to:
+            // - https://github.com/apache/iceberg/pull/4364
+            // - and deduplication feature (possible that old record isn't present in the table)
+            writer.deleteKey(row);
         }
+
         // if its deleted row and upsertKeepDeletes = true then add deleted record to target table
         // else deleted records are deleted from target table
-        if (
-                upsertKeepDeletes
-                        || !(row.getField("__op").equals(DELETE.getCode())))// anything which not an insert is upsert
-        {
+        if (upsertKeepDeletes || !isDelete) {
             writer.write(row);
         }
     }
 
-    public class RowDataDeltaWriter extends BaseEqualityDeltaWriter {
-        RowDataDeltaWriter(PartitionKey partition) {
+    public class RecordDeltaWriter extends BaseEqualityDeltaWriter {
+        RecordDeltaWriter(PartitionKey partition) {
             super(partition, schema, deleteSchema);
         }